refactor: NetworkErrorHandler 클래스로 중복되는 코드 제거
