+++ SOPT/SOPT/week4-project/HisHobby/HisHobbyService.swift
@@ -25,11 +25,10 @@ class HisHobbyService {
             headers: headers
         )
         .validate()
-        .responseData { [weak self] response in
+        .responseData { response in
             
             guard let statusCode = response.response?.statusCode,
-                  let data = response.data,
-                  let self
+                  let data = response.data
             else {
                 completion(.failure(.unknownError))
                 return
@@ -42,39 +41,14 @@ class HisHobbyService {
                     let hobby = hobbyResponse.result.hobby
                     completion(.success(hobby))
                 } catch {
-                    let error = self.handleStatusCode(statusCode, data: data)
+                    let error = NetworkErrorHandler.handleStatusCode(statusCode, data: data)
                     completion(.failure(error))
                 }
                 
             case .failure:
-                let error = self.handleStatusCode(statusCode, data: data)
+                let error = NetworkErrorHandler.handleStatusCode(statusCode, data: data)
                 completion(.failure(error))
             }
         }
-    }
-    
-    func handleStatusCode(
-        _ statusCode: Int,
-        data: Data
-    ) -> NetworkError {
-        let errorCode = decodeError(data: data)
-        switch (statusCode, errorCode) {
-        case (404, "00"):
-            return .invalidURL
-        case (404, "01"):
-            return .noNo
-        case (500, ""):
-            return .serverError
-        default:
-            return .unknownError
-        }
-    }
-    
-    func decodeError(data: Data) -> String {
-        guard let errorResponse = try? JSONDecoder().decode(
-            ErrorResponse.self,
-            from: data
-        ) else { return "" }
-        return errorResponse.code //00 혹은 01
     }
 }
